Added new code module class
